% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyfilesdialog-package.R
\docType{package}
\name{tinyfiledialogsR-package}
\alias{tinyfiledialogsR}
\alias{tinyfiledialogsR-package}
\title{tinyfiledialogsR: An R Package for Device Independent Dialogs}
\description{
This package in an R wrapper for the tinyfiledialogs C code, providing device independent dialogs. See the demo to explore the dialogs interactively.
}
\author{
\strong{Maintainer}: Paul Smith \email{paul@waternumbers.co.uk} (\href{https://orcid.org/0000-0002-0034-3412}{ORCID})

Authors:
\itemize{
  \item Guillaume Vareille \email{tinyfiledialogs@ysengrin.com}
}

}
\keyword{internal}
